Move UI Compose tests in screen or component subpackages
